requirements.txt
opencv-python==4.7.0.68
numpy==1.24.2
pandas==1.5.3
mediapipe-silicon==0.9.1